Eliminate some unnecessary allocations and `to_string()`
What can be taken by reference, is taken by reference.
This gives `Request` a lifetime, so it can't derive `Deserialize`.

Because of that, I had to delete a test, but that seems OK, since
it looks like this test is testing `serde_json` crate, and not any
logic in this crate.