Implement a basic "TCP" transport
The modularization of the HTTP-specific code allowed us to plug more
transports to the crate. This a first additional one, a pretty basic and
synchronous transport over a TCP listener.

The module is strongly inspired by Steven Roose's implementation of the
simple_http transport. Apart that the struct is so simple i did not
implemented a builder (seems way overkill).

I called it 'TCP' transport by lack of other name, but it does not
support TCP over Unix Domain Sockets. Another transport (because of
Rust's complete separation of both) will be added to support those.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>