lib: use serde_json::Value for the Request's params field
The specs explicitly don't restrict the type of the 'params' field:
> A Structured value that holds the parameter values to be used during the invocation of the method.
(src: https://www.jsonrpc.org/specification)

However, it's followed by:
> This member MAY be omitted.

But i didn't go as far as making it an Option.

This also makes the 'method' and 'jsonrpc' fields a Cow, as serde needs
to modify it during the serialization. An alternative if we want to
keep them as str would be to have a custom Deserialize impl
(by deserializing first to a String).

Also, using a serde_json::Value allows us to derive Deserialize!