Ground Work: Refactor URL check from builder api
To use the same url code checking for proxy address, the current
`builder::url()` function is broken down into two pieces. The first part
which checks the url sanity lives on it's own independent function
as `check_url()`.

This allows us to reuse same sanity check in to different function.