simple_http:  handle socket returning "connection reset on peer" errors
On Mac Os it was observed that a broken socket could return the error "Connection reset by peer" on the attempt to read or write when the socket was broken. This commit won't propagate those errors on first attempt and will ask for a fresh socket once to try again before returning an error.