minreq_http: return an HTTP error on error with no JSON in body
It is useful for downstream users to be matching on errors that do not
contain a valid JSONRPC error in the HTTP response body. One instance is
if the HTTP server work queue depth is exceeded, as they probably want
to retry the request later.

On such error we would return a JSON deserialization error, without
exposing neither the HTTP status code nor the body of the response. This
made it impossible to detect such transient errors.

Instead, introduce a new HttpError variant that gets returned when the
requested is responded to by an error that does not contain valid JSON
in its body.